backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
aiohttp
requests
beautifulsoup4
starlette
anyio
idna
sniffio
colorama
httptools
PyYAML
watchfiles
websockets
python-dateutil
fuzzywuzzy
python-Levenshtein  # Optional but recommended for better performance
selenium
psutil  # For process management and cleanup
pywin32  # For Windows signal handling and process management
numpy  # For numerical operations
sqlalchemy  # For database operations 